recover dydx v3 & reduce dydx_v4_auth
