+++ hummingbot/connector/derivative/dydx_perpetual/dydx_perpetual_user_stream_data_source.py
@@ -12,6 +12,7 @@ from hummingbot.logger import HummingbotLogger
 
 
 class DydxPerpetualUserStreamDataSource(UserStreamTrackerDataSource):
+
     _logger: Optional[HummingbotLogger] = None
 
     @classmethod
@@ -20,12 +21,10 @@ class DydxPerpetualUserStreamDataSource(UserStreamTrackerDataSource):
             cls._logger = logging.getLogger(__name__)
         return cls._logger
 
-    def __init__(self, dydx_auth: DydxPerpetualAuth, api_factory: Optional[WebAssistantsFactory], connector):
+    def __init__(self, dydx_auth: DydxPerpetualAuth, api_factory: Optional[WebAssistantsFactory]):
         self._dydx_auth: DydxPerpetualAuth = dydx_auth
         self._api_factory: WebAssistantsFactory = api_factory
         self._ws_assistant: Optional[WSAssistant] = None
-        self._connector = connector
-
         super().__init__()
 
     @property
@@ -43,21 +42,17 @@ class DydxPerpetualUserStreamDataSource(UserStreamTrackerDataSource):
             self._ws_assistant = await self._api_factory.get_ws_assistant()
             await self._ws_assistant.connect(ws_url=CONSTANTS.DYDX_WS_URL, ping_timeout=CONSTANTS.HEARTBEAT_INTERVAL)
 
-            subaccount_id = f"{self._connector._dydx_perpetual_chain_address}/{self._connector.subaccount_id}"
+            auth_params = {
+                "type": "subscribe",
+                "channel": CONSTANTS.WS_CHANNEL_ACCOUNTS,
+                "accountNumber": "0",
+            }
 
-            subscribe_account_request: WSJSONRequest = WSJSONRequest(
-                payload={
-                    "type": "subscribe",
-                    "channel": CONSTANTS.WS_CHANNEL_ACCOUNTS,
-                    "id": subaccount_id,
-                },
-                is_auth_required=False,
-            )
-            await self._ws_assistant.send(subscribe_account_request)
+            auth_request: WSJSONRequest = WSJSONRequest(payload=auth_params, is_auth_required=True)
+            await self._ws_assistant.send(auth_request)
             self.logger().info("Authenticated user stream...")
         return self._ws_assistant
 
     async def _process_event_message(self, event_message: Dict[str, Any], queue: asyncio.Queue):
         if event_message.get("type", "") in [CONSTANTS.WS_TYPE_SUBSCRIBED, CONSTANTS.WS_TYPE_CHANNEL_DATA]:
             await super()._process_event_message(event_message=event_message, queue=queue)
-
