+++ hummingbot/connector/derivative/dydx_perpetual/dydx_perpetual_web_utils.py
@@ -5,20 +5,8 @@ from hummingbot.connector.time_synchronizer import TimeSynchronizer
 from hummingbot.connector.utils import TimeSynchronizerRESTPreProcessor
 from hummingbot.core.api_throttler.async_throttler import AsyncThrottler
 from hummingbot.core.web_assistant.auth import AuthBase
-from hummingbot.core.web_assistant.connections.data_types import RESTMethod, RESTRequest
-from hummingbot.core.web_assistant.rest_pre_processors import RESTPreProcessorBase
+from hummingbot.core.web_assistant.connections.data_types import RESTMethod
 from hummingbot.core.web_assistant.web_assistants_factory import WebAssistantsFactory
-
-
-class DydxPerpetualRESTPreProcessor(RESTPreProcessorBase):
-
-    async def pre_process(self, request: RESTRequest) -> RESTRequest:
-        if request.headers is None:
-            request.headers = {}
-        request.headers["Accept"] = (
-            "application/json"
-        )
-        return request
 
 
 def public_rest_url(path_url: str, domain: str = CONSTANTS.DEFAULT_DOMAIN) -> str:
@@ -42,26 +30,20 @@ def private_rest_url(path_url: str, domain: str = CONSTANTS.DEFAULT_DOMAIN) -> s
 
 
 def build_api_factory(
-        throttler: AsyncThrottler,
-        # time_synchronizer: TimeSynchronizer,
-        auth: AuthBase,
-        time_provider: Optional[Callable] = None,
+    throttler: AsyncThrottler,
+    time_synchronizer: TimeSynchronizer,
+    auth: AuthBase,
+    time_provider: Optional[Callable] = None,
 ) -> WebAssistantsFactory:
-    throttler = throttler or create_throttler()
-    # time_provider = time_provider or (lambda: get_current_server_time(throttler=throttler))
+    time_provider = time_provider or (lambda: get_current_server_time(throttler=throttler))
     api_factory = WebAssistantsFactory(
         throttler=throttler,
         auth=auth,
         rest_pre_processors=[
-            DydxPerpetualRESTPreProcessor(),
-            # TimeSynchronizerRESTPreProcessor(synchronizer=time_synchronizer, time_provider=time_provider),
+            TimeSynchronizerRESTPreProcessor(synchronizer=time_synchronizer, time_provider=time_provider),
         ],
     )
     return api_factory
-
-
-def create_throttler() -> AsyncThrottler:
-    return AsyncThrottler(CONSTANTS.RATE_LIMITS)
 
 
 async def get_current_server_time(throttler: AsyncThrottler, domain: str = CONSTANTS.DEFAULT_DOMAIN) -> float:
