+++ hummingbot/connector/derivative/dydx_v4_perpetual/dydx_v4_perpetual_web_utils.py
@@ -1,22 +1,34 @@
 from typing import Callable, Optional
 
-import hummingbot.connector.derivative.dydx_perpetual.dydx_perpetual_constants as CONSTANTS
+import hummingbot.connector.derivative.dydx_v4_perpetual.dydx_v4_perpetual_constants as CONSTANTS
 from hummingbot.connector.time_synchronizer import TimeSynchronizer
 from hummingbot.connector.utils import TimeSynchronizerRESTPreProcessor
 from hummingbot.core.api_throttler.async_throttler import AsyncThrottler
 from hummingbot.core.web_assistant.auth import AuthBase
-from hummingbot.core.web_assistant.connections.data_types import RESTMethod
+from hummingbot.core.web_assistant.connections.data_types import RESTMethod, RESTRequest
+from hummingbot.core.web_assistant.rest_pre_processors import RESTPreProcessorBase
 from hummingbot.core.web_assistant.web_assistants_factory import WebAssistantsFactory
+
+
+class DydxV4PerpetualRESTPreProcessor(RESTPreProcessorBase):
+
+    async def pre_process(self, request: RESTRequest) -> RESTRequest:
+        if request.headers is None:
+            request.headers = {}
+        request.headers["Accept"] = (
+            "application/json"
+        )
+        return request
 
 
 def public_rest_url(path_url: str, domain: str = CONSTANTS.DEFAULT_DOMAIN) -> str:
     """
     Creates a full URL for provided public REST endpoint
     :param path_url: a public REST endpoint
-    :param domain: the dYdX domain to connect to ("exchange" or "us"). The default value is "exchange"
+    :param domain: the dydx_v4 domain to connect to ("exchange" or "us"). The default value is "exchange"
     :return: the full URL to the endpoint
     """
-    return CONSTANTS.DYDX_REST_URL + path_url
+    return CONSTANTS.DYDX_V4_REST_URL + path_url
 
 
 def private_rest_url(path_url: str, domain: str = CONSTANTS.DEFAULT_DOMAIN) -> str:
@@ -26,24 +38,26 @@ def private_rest_url(path_url: str, domain: str = CONSTANTS.DEFAULT_DOMAIN) -> s
     :param domain: the dYdX domain to connect to ("exchange" or "us"). The default value is "exchange"
     :return: the full URL to the endpoint
     """
-    return CONSTANTS.DYDX_REST_URL + path_url
+    return CONSTANTS.DYDX_V4_REST_URL + path_url
 
 
 def build_api_factory(
-    throttler: AsyncThrottler,
-    time_synchronizer: TimeSynchronizer,
-    auth: AuthBase,
-    time_provider: Optional[Callable] = None,
+        throttler: AsyncThrottler,
 ) -> WebAssistantsFactory:
-    time_provider = time_provider or (lambda: get_current_server_time(throttler=throttler))
+    throttler = throttler or create_throttler()
+    # time_provider = time_provider or (lambda: get_current_server_time(throttler=throttler))
     api_factory = WebAssistantsFactory(
         throttler=throttler,
-        auth=auth,
         rest_pre_processors=[
-            TimeSynchronizerRESTPreProcessor(synchronizer=time_synchronizer, time_provider=time_provider),
+            DydxV4PerpetualRESTPreProcessor(),
+            # TimeSynchronizerRESTPreProcessor(synchronizer=time_synchronizer, time_provider=time_provider),
         ],
     )
     return api_factory
+
+
+def create_throttler() -> AsyncThrottler:
+    return AsyncThrottler(CONSTANTS.RATE_LIMITS)
 
 
 async def get_current_server_time(throttler: AsyncThrottler, domain: str = CONSTANTS.DEFAULT_DOMAIN) -> float:
